Modificato GestioneOperazioni

+++ OperazioneMatematicaSpringWebMVC/src/main/java/it/dstech/controller/GestioneOperazioni.java
@@ -2,11 +2,11 @@ package it.dstech.controller;
 
 public class GestioneOperazioni {
 	
-	public int somma(int num1, int num2) {		
+	public double somma(int num1, int num2) {		
 		return num1+num2;
 	}
 	
-	public int sottrazione(int num1, int num2) {		
+	public double sottrazione(int num1, int num2) {		
 		return num1-num2;
 	}
